refactor: 인수테스트 픽스처 이름 변경 fixtures -> fixture
