+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/fixtures/CommentFixture.java
@@ -15,7 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class CommentFixtures {
+public class CommentFixture {
 
     private static ExtractableResponse<Response> 댓글을_등록한다(AccessTokenResponse tokenResponse,
                                                           ArticleIdResponse articleIdResponse,
